Set scope from vim var g:ncm2_html_subscope_scope
The set of filetypes recognized as html files is hardcoded into the
plugin. As a result, when there is a new html filetype, the new filetype
is not recognized by the plugin without having to manually change the
plugin itself to include the new html filetype in the set of recognized
filetypes.

This change allows setting a list of html filetypes from vim
configuration via a new global variable `g:ncm2_html_subscope_scope`. If
this global variable does not exist, then the default list of html
filetypes that's hardcoded into the plugin is used instead, and
`g:ncm2_html_subscope_scope` is also set to the default list of html
filetypes.